Improve bootstrap-contextmenu knockout binding.
To get at the actual DOM element clicked, we can use the 'scopes'
feature which specifies which element to pass to the onItem callback.
From there we can take advantage of ko.dataFor to get at the bound
node (TreeNode in this case).

This removes the need for the ad-hoc elementFromPoint logic in
the before handler.