Add support for multi-file refactorings.
Instead of just printing the new text to stdout or an error to
stderr, we print a JSON response, which has potentially three
keys:

exception: The stack trace if something went wrong (for debugging)
errors: An array of error messages (e.g. precondition failed)
modified: An object mapping affected filenames to new contents

The only change in McLab is the addition of a getModifiedPrograms()
method on the refactoring context. As an approximation, we consider
a program modified if its Transformer was accessed.