Add Octave support.
A new "backend" setting can be set to either "MATLAB" or "Octave";
this affects both the shell and the call graph instrumentation.

This depends on arokem/python-matlab-bridge#75.

A sort-of-unrelated change in this commit is that the Matlab or
Octave session is initialized asynchronously after the page has
loaded, instead of making the page hang.