Skip instrumenting DHO and DWH too, not just DO
Also fix a bug where mightBeFunctionHandle was not taken into account
in the second pass.

Also, use the out set of the handle propagation to make the decision to
instrument. Consider something like

a(i) = a(i) + 3

Here a is definitely not a handle -- or rather, if it is, we'll get a
runtime error. But if a is a function parameter, the value for a in the
set will be {H, DWH}, so we'll instrument it anyway. In the out set a
will be {DO}, so use that instead.