Replace loop eval with cell array, not matrices.
Anything you can assign to a variable, you can assign to a cell array
index. With matrices, the dimensions of the index and the rhs expression
have to match -- with a scalar index like a loop variable, only scalar
expressions can be assigned. So cell arrays are a safer choice.